CategoryItem type
