+++ linx/utils/types/newsletterJSON.ts
@@ -1,4 +1,5 @@
 import { NavigationInfo } from "./shared-types/navigations.ts";
+import { ResponseCallBack } from "./shared-types/responseCallBack.ts";
 
 export interface NewsletterResponse {
   Response: Response;
@@ -41,10 +42,4 @@ export interface Response {
   SuccessMessage: null;
   Url: null;
   Warnings: unknown[];
-}
-
-export interface ResponseCallBack {
-  Code: string;
-  Parameters: unknown[];
-  Value: string;
 }
