+++ linx/utils/types/productJSON.ts
@@ -3,8 +3,10 @@
  */
 
 import { LinxError } from "./common.ts";
+import { CategoryItem } from "./shared-types/categoryItem.ts";
 import { LinxMetadata } from "./shared-types/linxMetada.ts";
 import { NavigationInfo } from "./shared-types/navigations.ts";
+import { ProductSelectionItem } from "./shared-types/productSelection.ts";
 import { ResponseCallBack } from "./shared-types/responseCallBack.ts";
 
 export interface WebPage {
@@ -85,7 +87,7 @@ export interface Model {
   ProductDefinition: ProductDefinition;
   ProductDefinitionID: number;
   ProductID: number;
-  ProductSelection: ProductSelection;
+  ProductSelection: ProductSelectionItem;
   ProductTypeID: number;
   PromotionFrom: null;
   PromotionOnlyCheckout: boolean;
@@ -116,12 +118,6 @@ export interface Model {
 
   VisibleFrom: string;
   VisibleTo: null;
-}
-
-export interface CategoryItem {
-  ID: string;
-  Name: string;
-  ParentID: string;
 }
 
 export interface Item {
@@ -300,12 +296,6 @@ export interface ProductDefinition {
   ProductDefinitionID: number;
 }
 
-export interface ProductSelection {
-  PropertyLevel: number;
-  PropertyPath: string;
-  Quantity: number;
-  SkuID: number;
-}
 
 export interface Tag {
   Label: string;
