--- linx/utils/types/shared-types/webPage.ts
@@ -1,6 +1,0 @@
-export interface WebPageCore<T> {
-  Response: T;
-  Model: T;
-  PageInfo: T;
-  PageParamenters: T;
-}
